requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.9
dnspython==2.7.0
feedparser==6.0.11
Flask==3.1.0
Flask-PyMongo==2.3.0
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pymongo==4.3.3
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
reader==3.16
requests==2.32.3
scikit-learn==1.5.2
scikit-surprise==1.1.4
scipy==1.14.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
